refactor: request 패키지 생성
